+++ app/src/main/java/uk/ac/york/nimblefitness/Screens/RoutineAndExercise/EndSummaryFragment.java
@@ -73,7 +73,7 @@ public class EndSummaryFragment extends Fragment {
         int routineTotalMoves = 0;
 
         for(int i = 0; i < routine.getExerciseArrayList().size(); i++){
-            routineTotalMoves = routineTotalMoves + routine.getExerciseArrayList().get(i).getMovesPerRep()*routine.getExerciseArrayList().get(i).getReps();
+            routineTotalMoves = (int) (routineTotalMoves + routine.getExerciseArrayList().get(i).getMovesPerRep()*routine.getExerciseArrayList().get(i).getReps());
         }
 
         routineTotalMoves = routineTotalMoves*routine.getSets();
